Separated pawn movement validation to a separate file
